Rename test2 function
